+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API/UseCases/Reports/AllProjectsReportService.cs
@@ -33,7 +33,10 @@ using Dfe.ManageFreeSchoolProjects.API.UseCases.Project.Tasks.MovingToOpen;
 using Dfe.ManageFreeSchoolProjects.API.UseCases.Project.Tasks.PupilNumbersChecks;
 using Dfe.ManageFreeSchoolProjects.API.UseCases.Project.Tasks.PDG;
 using Dfe.ManageFreeSchoolProjects.API.UseCases.Project.Tasks.PrincipalDesignate;
+using Dfe.ManageFreeSchoolProjects.API.UseCases.Project.Tasks.ReadinessToOpenMeeting;
 using Dfe.ManageFreeSchoolProjects.API.UseCases.Project.Tasks.ReferenceNumbers;
+using Dfe.ManageFreeSchoolProjects.API.UseCases.Project.Tasks.PreFundingAgreementCheckpointMeeting;
+
 
 namespace Dfe.ManageFreeSchoolProjects.API.UseCases.Reports
 {
@@ -138,6 +141,7 @@ namespace Dfe.ManageFreeSchoolProjects.API.UseCases.Reports
                                       RiskAppraisalMeeting = RiskAppraisalMeetingTaskBuilder.Build(riskAppraisalMeetingTask),
                                       PDGDashboard = PDGDashboardBuilder.Build(po),
                                       KickOffMeeting = KickOffMeetingTaskBuilder.Build(kpi, milestones),
+                                      PreFundingAgreementCheckpointMeetingTask = PreFundingAgreementCheckpointMeetingTaskBuilder.Build(milestones),
                                       FundingAgreement = FundingAgreementTaskBuilder.Build(milestones),
                                       FundingAgreementHealthCheck = FundingAgreementHealthCheckTaskBuilder.Build(milestones),
                                       FundingAgreementSubmission = FundingAgreementSubmissionTaskBuilder.Build(milestones),
@@ -158,7 +162,8 @@ namespace Dfe.ManageFreeSchoolProjects.API.UseCases.Reports
                                       CommissionedExternalExpert = CommissionedExternalExpertTaskBuilder.Build(milestones),
                                       MovingToOpen = MovingToOpenTaskBuilder.Build(kpi,milestones),
                                       PrincipalDesignate = PrincipalDesignateTaskBuilder.Build(milestones),
-                                      DueDiligenceChecks = DueDiligenceChecksTaskBuilder.Build(milestones)
+                                      DueDiligenceChecks = DueDiligenceChecksTaskBuilder.Build(milestones), 
+                                      ReadinessToOpenMeetingTask = ROMTaskBuilder.Build(milestones)
                                   },
                                   ProjectReferenceData = new ProjectReferenceData()
                                   {