+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects/Constants/RouteConstants.cs
@@ -172,6 +172,9 @@ namespace Dfe.ManageFreeSchoolProjects.Constants
         
         public const string ViewDueDiligenceChecks = TaskList + "/due-diligence-checks";
         public const string EditDueDiligenceChecks = ViewDueDiligenceChecks + "/edit";
+
+        public const string ViewPreFundingAgreementCheckpointMeeting = TaskList + "/pre-funding-agreement-checkpoint-meeting";
+        public const string EditPreFundingAgreementCheckpointMeeting = ViewPreFundingAgreementCheckpointMeeting + "/edit";
         
         public const string ViewReadinessToOpenMeeting = TaskList + "/readiness-to-open-meeting";
         public const string EditReadinessToOpenMeeting = ViewReadinessToOpenMeeting + "/edit";